+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - [**Feature**] Restart dothttp-cli command
 - [**Improvement**] save file before running (for http file, for notebook it will not try to save)
 - [**Bug**] import postman shows error message in case of failure
+- [**Feature**] supports vscode hovers(shows informational message on hovering in dothttp)
+- [**Feature**] supports vscode definition(on clicking name, it will direct you to base definition)
 - [**Improvement**] Bump dothttp-cli to 0.0.24
   - [**Improvement**] import postman now supports 2.1.0 postman collection (also lot of bug fixes with import)